Refactor SliderValueLabel
