+++ packages/base/Slider/src/SliderValueLabel/SliderValueLabel.tsx
@@ -1,17 +1,9 @@
 import type { SliderValueLabelSlotProps } from '@mui/base/Slider'
-import React, { useEffect, useRef } from 'react'
-import { twJoin } from 'tailwind-merge'
+import React from 'react'
 
-import { useSliderContext } from '../Slider/SliderContext'
-import { getPosition } from '../slider-utils/slider-utils'
+import useSliderValueLabel from './useSliderValueLabel'
 
 type ValueLabelDisplay = 'on' | 'auto' | 'off'
-
-const tooltipStates: Record<ValueLabelDisplay, string> = {
-  off: 'hidden',
-  auto: 'hidden group-hover/thumb:flex justify-center items-center',
-  on: 'flex justify-center items-center',
-}
 
 const SliderValueLabel = ({
   children,
@@ -23,35 +15,15 @@ const SliderValueLabel = ({
   tooltip: ValueLabelDisplay
   isOnScreen: boolean
 }) => {
-  const { registerValueLabel, hasTooltipOverflow, checkTooltipsOverlap } =
-    useSliderContext()
-  const ref = useRef<HTMLSpanElement>(null)
-  const sliderValue = ownerState.value
-  const isRangeSlider = Array.isArray(sliderValue)
-  const isRangeSliderCollapsed =
-    isRangeSlider && sliderValue[0] === sliderValue[1]
-
-  useEffect(() => {
-    if (!isRangeSlider) {
-      return
-    }
-    checkTooltipsOverlap()
-  }, [checkTooltipsOverlap, sliderValue, isRangeSlider])
-
-  useEffect(() => {
-    registerValueLabel(index, ref)
-  }, [index, registerValueLabel])
+  const { ref, classes } = useSliderValueLabel({
+    index,
+    tooltip,
+    isOnScreen,
+    ownerState,
+  })
 
   return (
-    <span
-      ref={ref}
-      className={twJoin(
-        'absolute will-change-transform shadow-4 transition-transform m-1 text-sm bg-graphite-800 text-white rounded-sm py-[2px] px-2',
-        tooltipStates[tooltip],
-        isOnScreen ? 'bottom-[calc(100%+4px)]' : 'top-[calc(100%+4px)]',
-        getPosition({ hasTooltipOverflow, isRangeSliderCollapsed, index })
-      )}
-    >
+    <span ref={ref} className={classes}>
       {children}
     </span>
   )
